chore: move thread_pool.py to tools
